extract select source fields
